+++ src/main/java/org/terasology/polyworld/graph/GraphFacetProvider.java
@@ -19,12 +19,12 @@ import org.terasology.engine.world.generation.Facet;
 import org.terasology.engine.world.generation.GeneratingRegion;
 import org.terasology.engine.world.generation.Produces;
 import org.terasology.engine.world.generation.Requires;
-import org.terasology.math.delaunay.Voronoi;
 import org.terasology.math.geom.Rect2f;
 import org.terasology.math.geom.Rect2i;
 import org.terasology.math.geom.Vector2f;
 import org.terasology.nui.properties.Range;
 import org.terasology.polyworld.TriangleLookup;
+import org.terasology.polyworld.math.delaunay.Voronoi;
 import org.terasology.polyworld.rp.RegionType;
 import org.terasology.polyworld.rp.WorldRegion;
 import org.terasology.polyworld.rp.WorldRegionFacet;
