+++ src/main/java/org/terasology/polyworld/math/delaunay/ICoord.java
@@ -1,0 +1,11 @@
+// Copyright 2020 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
+
+package org.terasology.polyworld.math.delaunay;
+
+import org.terasology.math.geom.Vector2f;
+
+public interface ICoord {
+
+    Vector2f getCoord();
+}
